+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 Change Log
 ==========
 
+Version 1.1.0 *(2013-04-16)*
+----------------------------
+* Added "start" and "end" buttons
+* Added new demo "Buttons in landscape mode only"
+
 
 Version 1.0.0 *(2013-04-09)*
 ----------------------------
-Initial release.
+* Initial release.